:smile_cat:
Fixed the encrypt_file() routine, this now returns the correct data, and has an HMAC of sha512 built in as the Tag itself. read the code to figure out how its implemented so that way you can extend it if you want. size is an issue, the files are currently 400 KB, need to figure out a way to condense that down, and still provide the features needed within this shell.