:smile_cat:
corrected a few issues in the self-destruct function. this will encrypt the file, unlink it, and attempt to cause a crash on the machine by hogging system resources if it is identified to be for reverse engineering. Going to add in a way to have the script call back with an hmac of the dropped file/shell to ensure no tampering has occured, with 1 failure causing a self destruct.