:smile_cat:
corrected a few issues. The call to initially connect to the database is now properly set up to the postgres user, once that fires, the application should connect with the needed user.  Corrected a few issues last night regarding the client script, and issues with it doing bulk checking if hosts are still owned or not, going to add something in the menu that allows for listing/query of owned hosts.