:smile_cat:
Will be adding in function to update proxies from proxyscrape. This will also include a new column in the db for highest viable/someother method. The proxy rotation function will need some fine tuning and further developments included. Will also be taking a look at the encrypt command when sending commands and why the shell on the victim is no longer able to decrypt.