:smile_cat:
corrected a few things with this, slimmed down the dropper, will find a way to work in the obfuscation routine to the dropper. adding a way for diag_handler to kick out a base64 encoded version of the actual slop shell to be written to disk, but for now.... this is it.