:smile_cat:
Added a few fixes to the new dropper creator. This is quite fast, maximum length of the entire dropper is 1100 bytes in length. All droppers will return a 404 no matter what you end up doing, the only difference you will see, is you will get a hit on the diag_handler.php script with the variable r set to pull. Going to make sure this is more stealthy, through cookies.