:smile_cat:
Setting a few values and correcting a few tell tales that will give the slimmed down dropper away. So far its just under 1.3kb (had to step up the size to ensure security of the dropper and shell. Moved the main line of communication between operator and dropper to cookies only, since most cookies are not logged. May end up adding encryption to the cookie value itself to ensure that even if this is caught, that the chance of the C2 being caught right off of the gate/entry is slim to almost none. Since this shell leverages the strength of PGSQL, all relative data to the dropper is stored in the db for retrieval when needed, can even slim this down to reference the planned target so no 2 shells will be mixed up. However, since the value is serialized it should be relatively easy to swap the delivery host if needed.