:smile_cat:
added a randomized UA with a random string appended to the end. this will allow for more stealth when activating the dropper. Going to add a few functions in the client script to activate the dropper, and then reset it. Trying to keep these php scripts in memory so they dont hit disk.