:smile_cat:
Added a check on tor, to verify its running. In the event the user chooses tor as their proxy. corrected a bug in the update_proxy function, this will no longer cause a crash. working to fix the encryption of commands. trying to use a pure php method. we shall see.