:smile_cat:
Adding host enumeration script as a generic tool to be used by this framework.

This will be encrypted C2 side, requiring an X509/Public key to decrypt, not every key will be the same so many will be generated or generated over time. Each key will be different by agent(making tracking movement more difficult.)

The base functionality of this script, is it enumerates all php functions and extensions loaded by the php env, current processes running, and current users on the system linting out the ones who are of no use.