:smile_cat:
I LIED ON MY LAST ONE. THIS IS THE LAST COMMIT. DONT RUN THIS WITH THE CLIENT SCRIPT, IT WILL NOT BE ABLE TO COMMUNICATE USING THE NEW SCHEME. 

The file itself will delete itself once execution is done, As you can see from the code, the new file is in the header aptly named NewName. This will serve as a rolling way to hide the script, as the server admin will likely look for the one showing up in the logs, and not the new filename.