:smile_cat:
Added a way to pack the whole slop.php trojan into a wordpress plugin, with init hook. Need to test and see if this can be ignored(i know it can.) and still hook the website/wordpress install. Still working through ways to add a randomized slopp function name, but it appears as though php has no straight forward way to replace an array index completely. Still working through the rest of the script to tie all new features into the client script.