:smile_cat:
working on MTLS implementation of the client and the shell, this will result in an encrypted file being hidden on the FS somewhere, de-crypted, and then executed when the shell is called successfully. Working on methods to control where the contents of the shell will be delivered from, currently in works to have the shell sent via the server in each request, although this may bog down the whole process and speed is needed. will look to optimize this while process.