:smile_cat:
adding the ability to check for and start the tor binary, gonna need to work out the specific kinks in this, and make sure it can be done.

final step will be to daemonize this script and have it run forever, making it a proper trojan.

Need to see how i can protect the process from tampering, making it an unkillable process, might be a pipe dream, but im still curious to see.