:smile_cat:
simplified a few functions. moving all communications to use the HEAD request. less suspicious in logs, specially if there is a 404 response to the request. will work to compress the responses, so the size is smaller as a whole. Need to look into access log functions, to see if this can be hidden from the request log entirely.