:smile_cat:
Added pagination of information in the database.  Client will successfully execute commands and process the output in the users terminal. This will be expanded with more functionality, and the pagination will ALSO be expanded to have more functionality, right now its buggy and is prone if you exceed the last page to not print any results. oops. it happens. Will likely improve upon the pagination portion, just wanted to get it working.