:smile_cat:
added some hints for command types and encoding/encryption types that can be done from the client. right now the default action you need to take is 1b or 1. These are already worked in the shell and seamlessly flow. The decryption portion of the client has not been done yet. (work in progress.)