:smile_cat:
added a line to the client that now updates the entry in the db for the new name if the bot is able to do so. Will add checks in, or just leave it there and have it default to something else. Adding more stuff to it. Happy new year.