:smile_cat:
noticed the com function was entirely cumbersome. sometimes youll just want to continue command execution on a host. Added a loop that will let you run commands on the bot to your hearts content.

Working on adding in the raw sha1 sums of the required information.

And determining if i should use a HEAD request to transfer file chunks across to the shell or not, so not to break the stealth given by the head requests and the associated 404 requests.