:smile_cat:
i stopped being lazy. added in the sha1 sums so that way the needed values are not evident in the script. making tracing who it was that actually used the script unknown from the logs perspective, if the shell is named something non-descript, the randomization of file name will need to be reworked so that the shell hides more in the filesystem.

Added some more useful protections to enumerate for on the affected system.