:smile_cat:
added in more refined base templates, with replacement markers. will generate more text blobs to throw in, and more configs for particlesJS(can add more JS frameworks so its more dynamic and not repeated, work in progress and i hate UI development.) so that each page is different. Working on a way to have everything stored locally so that external calls are not conducted.